moved contextCreator to seperate package
